requirements.txt
azure-functions
pgpy
azure-storage-blob
requests
pyOpenSSL
ndg-httpsclient
 